\name{LausenTab2}
\alias{LausenTab2}
\title{
Table 2 of Lausen and Schumacher (1992)
}
\description{
Table 2 of Lausen and Schumacher (1992)
}
\usage{
data(LausenTab2)
}
\references{
Lausen, B. and Schumacher, M. (1992), Maximally Selected Rank Statistics.
\emph{Biometrics} \bold{48}, 73--85
}
\keyword{datasets}
