\name{maxstat-internal}
\alias{intrank}
\alias{p2norm}
\alias{p2normG}
\alias{pbvmax}

\title{Internal maxstat functions}
\description{
 Internal maxstat functions.
}
\usage{
intrank(x)
p2norm(h, k, rho, maxp=3000)
p2normG(h, k, rho, maxp=3000)
pbvmax(cor, t, method=c("drezner", "genz"))
}
\details{
  These functions are not to be called by the user.
}
\keyword{internal}
